test(protocol): de-duplicate noteregistry tests
